feat(query): Output format support.
This adjusts the output format according to what the UI requests. A new
`format` field has been introduced on the query payload that supports
the following values: `table` and `time_series`.

When `time_series` is chosen we do a long-to-wide conversion to create
the label series. `table` is just a direct passthrough of what we
receive from the upstream.