feat(client): Extract stream headers for query DoGet.
The API provided by the Flight SQL client doesn't allow access to gRPC
metadata for streaming operations such as `DoGet`.

To work around this problem, I've implemented a small wrapper for the
client that implements an additional `DoGetWithHeaderExtraction` method.
This method performs a normal `DoGet` (using the underlying Flight RPC
client), but returns a custom type that exposes the headers obtained
after the first stream read.

The headers are then applied to the outgoing dataframe's metadata, and
can be accessed using the Grafana query debugger.

We should consider opening a discussion with the upstream project about
improving the Flight SQL client to make header/trailer access possible.